Updated ContactRecords.proto, based on what Google published at https://github.com/google/exposure-notifications-internals/blob/main/exposurenotification/src/main/proto/contact_record.proto Still missing: Handling of the non-deprecated fields (but I couldn't find them 'in the wild' yet). Contributes to #70.

+++ corona-warn-companion/src/main/java/org/tosl/coronawarncompanion/matchentries/MatchesRecyclerViewAdapter.java
@@ -316,7 +316,7 @@ public class MatchesRecyclerViewAdapter extends RecyclerView.Adapter<MatchesRecy
                 }
                 byte txPower = aem[1];
                 //Log.d(TAG, "TXPower: "+txPower+" dBm");
-                int rssi = (int) scanRecord.getRssi();
+                int rssi = scanRecord.getRssi();
                 //Log.d(TAG, "RSSI: "+rssi+" dBm");
                 int attenuation = txPower - rssi;
                 //Log.d(TAG, "Attenuation: "+attenuation+" dB");
